+++ composer.json
@@ -11,8 +11,8 @@
     }
   ],
   "require": {
-    "illuminate/support": "5.1.*",
-    "tecnick.com/tcpdf": "6.2.9"
+    "illuminate/support": "5.0.x",
+    "tecnick.com/tcpdf": "6.2.6"
   },
   "autoload": {
     "psr-4": {
