+++ src/Facades/Pdf.php
@@ -4,10 +4,15 @@ namespace Elibyy\TCPDF\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
-class TCPdf extends Facade
+/**
+ * Class Pdf
+ * @version 1.0
+ * @package Elibyy\TCPDF\Facades
+ */
+class Pdf extends Facade
 {
 	protected static function getFacadeAccessor()
 	{
-		return 'tcpdf';
+		return 'pdf';
 	}
 }
